TA dev kit: shared library: add -f argument to ln
When creating a symbolic link <uuid>.elf -> libname.so, use the -f (force)
argument so that the command won't fail if the link already exists. It is
a very common case: the first time the library is built and the link is
created. Then whenever some source file is modified, the .so is re-linked
but the symbolic link is there already.

Fixes: 01b8b5ce011d ("TA dev kit: when building a shared library, create symlink with UUID")
Signed-off-by: Jerome Forissier <jerome.forissier@linaro.org>
Reviewed-by: Jens Wiklander <jens.wiklander@linaro.org>