ta services: base for secure key services and TA API
SKS aims at providing PKCS#11 compliant services through a
trusted application operating as a secure service provider.

This is the first step. It defines the trusted application,
its entry points and an initial API.

The TA API is defined in sks_ta.h and should be exported to the client.
This commit introduces the first command: SKS_CMD_PING, used to check TA
availability and later retrieve design version information.

Signed-off-by: Etienne Carriere <etienne.carriere@linaro.org>