user_ta: should go for other TA stores on any load error
There seems to be an issue that if RPMB_FS is enabled in
OPTEE and TA is present in REE (normal file system), if
priority for secure storage TA is higher and RPMB
initialization fails, the error is returned and the
OPTEE doesn't goes to find the TA from REE TA store.

The issue is fixed by adding a 'continue' statement after
printing the respective error.

Signed-off-by: Sourabh Das <sourabhdas143@gmail.com>