Do not make in-tree TAs depend on a phony target
Commit cf903a62bc12 ("Add dependency on ta-dev-kit when building in-tree
TA") introduces a dependency of the in-tree TA object files on the phony
target 'ta_dev_kit'. The purpose was to make sure the dev kit files are
available when make starts building the TA. Unfortunately, this introduces
useless recompilation (the lines marked with >):

 $ make -s && make
   CHK     out/arm-plat-vexpress/conf.mk
   CHK     out/arm-plat-vexpress/include/generated/conf.h
   CHK     out/arm-plat-vexpress/conf.cmake
   CHK     out/arm-plat-vexpress/export-ta_arm32/mk/conf.mk
>  CC      out/arm-plat-vexpress/ta/avb/entry.o
>  CC      out/arm-plat-vexpress/ta/avb/user_ta_header.o
>  CPP     out/arm-plat-vexpress/ta/avb/ta.lds
>  LD      out/arm-plat-vexpress/ta/avb/023f8f1a-292a-432b-8fc4-de8471358067.elf
>  OBJDUMP out/arm-plat-vexpress/ta/avb/023f8f1a-292a-432b-8fc4-de8471358067.dmp
>  OBJCOPY out/arm-plat-vexpress/ta/avb/023f8f1a-292a-432b-8fc4-de8471358067.stripped.elf
>  SIGN    out/arm-plat-vexpress/ta/avb/023f8f1a-292a-432b-8fc4-de8471358067.ta

This happens because phony targets are always considered new by the make
program, so they'd rather not be used as dependencies. Instead, replace
'ta-dev-kit' by the actual list of all the files in the TA dev kit. Then,
the second make invocation will not rebuild anything:

 $ make -s && make
   CHK     out/arm-plat-vexpress/conf.mk
   CHK     out/arm-plat-vexpress/include/generated/conf.h
   CHK     out/arm-plat-vexpress/conf.cmake
   CHK     out/arm-plat-vexpress/export-ta_arm32/mk/conf.mk

Signed-off-by: Jerome Forissier <jerome.forissier@linaro.org>
Reviewed-by: Jens Wiklander <jens.wiklander@linaro.org>