README.md: do not mention the checkpatch license
checkpatch.pl being covered by the GPL is not a convincing reason for
not having a local copy of the script. Nothing in OP-TEE is a derivative
work of checkpatch, and a "mere aggregation" is even explicitly covered in
the GPL text. It isn't necessary to be concerned with the output of
checkpatch, since that is not included in OP-TEE, either.

Therefore, remove the sentence that mentions the checkpatch license.

Reported-by: David Brown <david.brown@linaro.org>
Signed-off-by: Jerome Forissier <jerome.forissier@linaro.org>
Acked-by: David Brown <david.brown@linaro.org>
Acked-by: Jens Wiklander <jens.wiklander@linaro.org>