Fix copying dirty values in copy_in_params()
If the OP-TEE driver from the rich OS specifies a message with a number
of params < TEE_NUM_PARAMS, copy_in_params() will copy in undefined
values from pt[i] (where i >= the number of params). This is because the
pt array is an uninitialized local value, and per the C99 standard
6.7.8:

If an object that has automatic storage duration is not initialized
explicitly, its value is indeterminate.

This change fixes this issue by clearing out the unused parts of pt.

Signed-off-by: Christopher Tam <godtamit@google.com>
Reviewed-by: Jens Wiklander <jens.wiklander@linaro.org>