auto-locate checkpatch
Try to locate checkpatch.pl in typical location(s) if environment
variable CHECKPATCH is not set.
The launch script first tries to find checkpatch.pl on the path,
next it checks typical locations for linux headers, finally it
tries to locate linux sources used for OP-TEE in with QEMU.
The first match is used as the checkpatch instance to be invoked.

Signed-off-by: Markus S. Wamser <markus.wamser@mixed-mode.de>
Reviewed-by: Jerome Forissier <jerome.forissier@linaro.org>