core: force read-only flag on .rodata.* sections
This commit fixes a warning with GCC 8.2 that did not occur with GCC 6.2:

 $ make out/arm-plat-vexpress/core/arch/arm/kernel/user_ta.o
   CHK     out/arm-plat-vexpress/conf.mk
   CHK     out/arm-plat-vexpress/include/generated/conf.h
   CHK     out/arm-plat-vexpress/core/include/generated/asm-defines.h
   CC      out/arm-plat-vexpress/core/arch/arm/kernel/user_ta.o
 {standard input}: Assembler messages:
 {standard input}:4087: Warning: setting incorrect section attributes for .rodata.__unpaged

The message is printed as the assembler processes this code fragment,
generated by the C compiler:

 .section        .rodata.__unpaged,"aw"

The older compiler (GCC 6.2) would generate instead:

 .section        .rodata.__unpaged,"a",%progbits

The problem with .rodata.__unpaged,"aw" is that the "w" (writeable) flag
is not consistent with the section name (.rodata.*), which by convention
is supposed to be read-only.

- The section name (".rodata.__unpaged") is given by our macro:
  __rodata_unpaged.
- The "w" flag is added by GCC, not sure why exactly. One reason [1] is
  when a relocatable binary is being generated and the structure contains
  relocatable data. But, we are not explicitly asking for a relocatable
  binary, so this might as well be a bug or counter-intuitive feature of
  the compiler.

Anyway, to avoid the warning, we need to fix the section flags. The
section type (%progbits) is optional, it is deduced from the section name
by default. %progbits indicates that the section contains data (i.e., is
not empty).

Link: [1] https://gcc.gnu.org/ml/gcc/2004-05/msg01016.html
Signed-off-by: Jerome Forissier <jerome.forissier@linaro.org>
Tested-by: Jerome Forissier <jerome.forissier@linaro.org> (QEMU)
Tested-by: Jerome Forissier <jerome.forissier@linaro.org> (HiKey960)
Acked-by: Jens Wiklander <jens.wiklander@linaro.org>
Acked-by: Etienne Carriere <etienne.carriere@linaro.org>