core: generic_boot: Add add_dt_overlay_fragment()
This patch adds a dt routine add_dt_overlay_fragment(). This purpose of
which is to encapsulate the dynamic FDT node entries OPTEE provides inside
of a

fragment@0 {
        target-path = "/";
        __overlay__ {
                /* OPTEE nodes go here */
        };
};

A subsequent set of patches will wrapper up existing dynamic OPTEE nodes

- /firmware/optee
- /psci
- /reserved/memory
      optee@0xaddress

Once done it will be possible for a DTB in memory to be populated
with OPTEE fragments and for a subsequent bootloader to merge the OPTEE
populated overlay into a DTB loaded by a later stage.

if CFG_EXTERNAL_DTB_OVERLAY is not defined then this code has no effect.

Signed-off-by: Bryan O'Donoghue <bryan.odonoghue@linaro.org>
Reviewed-by: Jerome Forissier <jerome.forissier@linaro.org>
Acked-by: Jens Wiklander <jens.wiklander@linaro.org>