core: generic_boot: Add helper add_dt_path_subnode()
A common pattern in the DT code is

offs = fdt_path_offset(fdt, "/");
if (offs < 0)
	return -1;
offs = fdt_add_subnode(fdt, offs, "newnode");
if (offs < 0)
	return -1;

as was pointed out by Jerome in a related PR to-do with adding overlays
this is a candidate for functional decomposition.

This patch adds the necessary helper function.

Signed-off-by: Bryan O'Donoghue <bryan.odonoghue@linaro.org>
Suggested-by: Jerome Forissier <jerome.forissier@linaro.org>