core: base TA store on scatter array
The TA store interface is now implemented using scatter array.

The priorities of the different TA store are also changed to use 2, 4
and 9 instead or 5, 9, 10 as they are now sorted alphabetically. The new
allocation makes it easier to stick something between "Secure Storage
TA" and the "REE" stores. It's doubtful that anyone would ever want to
make something higher priority than the "early TA" store, but just in
case a few numbers are reserved.

Acked-by: Etienne Carriere <etienne.carriere@linaro.org>
Tested-by: Jerome Forissier <jerome.forissier@linaro.org> (HiKey960)
Reviewed-by: Jerome Forissier <jerome.forissier@linaro.org>
Signed-off-by: Jens Wiklander <jens.wiklander@linaro.org>