ta: avb: RSA public key validation
Introduce support for validation of RSA public key, used to verify
vbmeta partition. By validation it means that the given public key
is trusted and can be used for verification of vbmeta image signature.
For additional details check validate_vbmeta_public_key in [1].

It is also required by AVB spec [1] that trusted key is stored in the
tamper-evident storage.

The key data provided from NW is serialized in this format:
+--------------+----------------------------+
| Name         | Number of bytes            |
+--------------+----------------------------+
| key_num_bits | 4 ("modulus size in bits") |
| n0inv        | 4 ("-1/n[0] (mod 2^32)")   |
| n            | key_num_bits / 8 (modulus) |
| rr           | key_num_bits / 8 (R^2)     |
+--------------+----------------------------+

So, for example, for 4096 bit(default one) key representation we need
1032 bytes
(4 + 4 + 512 + 512) to store these values.

Public exponent is assumed to always be 65537.

Link: [1] https://android.googlesource.com/platform/external/avb/+/master/README.md#Named-Persistent-Values

Signed-off-by: Igor Opaniuk <igor.opaniuk@linaro.org>