stm32mp1: clocks: DT for sources frequencies and RCC secure state
Get the frequency of the source clocks from the device tree.

Secure status state in the device tree defines whether RCC
subsystem shall be secure or not. If not secure, non-secure
world can access all clock interface and secure world cannot
play with clock in a safe way. Yet, DT allows such debug or
test configuration.

Note that most clock tree configuration is under to responsibility
of an earlier boot stage. Configuration of parenthood and related
divisors as well as configuration of PLLs found in the DT are
ignored.

Signed-off-by: Etienne Carriere <etienne.carriere@st.com>